@@ -1,48 +1,63 @@
 {
-  "name": "resume_online-web",
-  "version": "0.1.0",
-  "private": true,
+  "name": "uni-preset-vue",
+  "version": "0.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev:app": "uni -p app",
+    "dev:app-android": "uni -p app-android",
+    "dev:app-ios": "uni -p app-ios",
+    "dev:custom": "uni -p",
+    "dev:h5": "uni",
+    "dev:h5:ssr": "uni --ssr",
+    "dev:mp-alipay": "uni -p mp-alipay",
+    "dev:mp-baidu": "uni -p mp-baidu",
+    "dev:mp-kuaishou": "uni -p mp-kuaishou",
+    "dev:mp-lark": "uni -p mp-lark",
+    "dev:mp-qq": "uni -p mp-qq",
+    "dev:mp-toutiao": "uni -p mp-toutiao",
+    "dev:mp-weixin": "uni -p mp-weixin",
+    "dev:quickapp-webview": "uni -p quickapp-webview",
+    "dev:quickapp-webview-huawei": "uni -p quickapp-webview-huawei",
+    "dev:quickapp-webview-union": "uni -p quickapp-webview-union",
+    "build:app": "uni build -p app",
+    "build:app-android": "uni build -p app-android",
+    "build:app-ios": "uni build -p app-ios",
+    "build:custom": "uni build -p",
+    "build:h5": "uni build",
+    "build:h5:ssr": "uni build --ssr",
+    "build:mp-alipay": "uni build -p mp-alipay",
+    "build:mp-baidu": "uni build -p mp-baidu",
+    "build:mp-kuaishou": "uni build -p mp-kuaishou",
+    "build:mp-lark": "uni build -p mp-lark",
+    "build:mp-qq": "uni build -p mp-qq",
+    "build:mp-toutiao": "uni build -p mp-toutiao",
+    "build:mp-weixin": "uni build -p mp-weixin",
+    "build:quickapp-webview": "uni build -p quickapp-webview",
+    "build:quickapp-webview-huawei": "uni build -p quickapp-webview-huawei",
+    "build:quickapp-webview-union": "uni build -p quickapp-webview-union"
   },
   "dependencies": {
-    "axios": "^0.27.2",
-    "core-js": "^3.8.3",
-    "element-ui": "2.15.10",
-    "nprogress": "^0.2.0",
-    "vue": "^2.6.14",
-    "vue-router": "3",
+    "@dcloudio/uni-app": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-app-plus": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-components": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-h5": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-alipay": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-baidu": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-kuaishou": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-lark": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-qq": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-toutiao": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-mp-weixin": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-quickapp-webview": "3.0.0-alpha-3060420220922001",
+    "vue": "^3.2.37",
+    "vue-i18n": "^9.1.9",
     "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "@babel/eslint-parser"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@dcloudio/types": "^3.0.13",
+    "@dcloudio/uni-automator": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-cli-shared": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/uni-stacktracey": "3.0.0-alpha-3060420220922001",
+    "@dcloudio/vite-plugin-uni": "3.0.0-alpha-3060420220922001",
+    "vite": "^2.9.14"
+  }
 }
