@@ -1,28 +1,10 @@
-import Vue from 'vue'
-import App from './App.vue'
-import ElementUI from "element-ui";
-import "element-ui/lib/theme-chalk/index.css";
-import router from "./router";
-
-Vue.use(ElementUI);
-
 import {
-  Loading,
-  MessageBox,
-  Message,
-  Notification,
-} from "element-ui";
-Vue.prototype.$loading = Loading.service;
-Vue.prototype.$msgbox = MessageBox;
-Vue.prototype.$alert = MessageBox.alert;
-Vue.prototype.$confirm = MessageBox.confirm;
-Vue.prototype.$prompt = MessageBox.prompt;
-Vue.prototype.$notify = Notification;
-Vue.prototype.$message = Message;
-
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app')
+	createSSRApp
+} from "vue";
+import App from "./App.vue";
+export function createApp() {
+	const app = createSSRApp(App);
+	return {
+		app,
+	};
+}
