@@ -1,25 +1,17 @@
-<template>
-  <div id="app">
-    <router-view/>
-  </div>
-</template>
-
 <script>
-
 export default {
-  name: 'App',
-  components: {
-  }
+  onLaunch: function () {
+    console.log('App Launch')
+  },
+  onShow: function () {
+    console.log('App Show')
+  },
+  onHide: function () {
+    console.log('App Hide')
+  },
 }
 </script>
 
 <style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-}
+/*每个页面公共css */
 </style>
